Added files via upload
this function judges whether there is a fish at position [i,j]; if there is,  it will be randomly moved to a new neighbor position available and the age is increased. if the fish reaches the breeding age, 2 fishes will be produced in the original and new position with age 0. number -1 means no fish.