Initial conditions may vary
I got the desired figure when size = 20, breed_age_fish = 1.0, breed_age_shark = 2.2, and initial fish 0.5%, initial shark 0.02%, the figure is not so smooth, see "population.jpg". When I change to grid size =100(Present parameters in the program), I have not acquired the desired figure since many parameters to be changed accordingly. See "population1.jpg", it is not that so good.