Create gp2_5
Problem fixed. There seem to be 2 problems in the previous version. First, did not initialize fishmove and sharkmove before each time step. Second, in shark_hunt function when we erase fish, we should erase it on fishmove instead of directly on fish. And seems the amplitude and period is highly sensitive to breed age of two species.